backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv==1.0.1
pydantic==2.9.2
pydantic_settings==2.13.1

# LangGraph / LangChain
langgraph
langchain
langchain-anthropic
langchain-openai

# Anthropic & OpenAI
anthropic
openai

# Supabase
supabase

# Voice
deepgram-sdk
azure-cognitiveservices-speech

# Notifications
python-telegram-bot

# MCP
mcp==1.26.0

# Auth / JWT
PyJWT==2.11.0
cryptography==46.0.5

# CORS / HTTP
httpx==0.28.1
python-multipart==0.0.22
